requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
apscheduler==3.10.4
requests==2.31.0
markdown==3.5.2 